+++ src/engine_core/resources/src/ResourceManager.cpp
@@ -1,5 +1,22 @@
-/*! \file ResourceManager.cpp
-	\author Alan Ramirez
-	\date 2024-12-24
-	\brief Resource manager
-*/
+#include "ResourceManager.hpp"
+
+Hush::RefCounted* Hush::ResourceManager::IncreaseRefCount(const HandleId& handle) {
+	RefCounted& count = this->m_references[handle];
+	count.count++;
+	return &count;
+}
+
+
+void Hush::ResourceManager::DecreaseRefCount(const HandleId& handle) {
+	RefCounted& count = this->m_references[handle];
+	// TODO: Add condition to prevent overflow
+	count.count--;
+	if (count.count == 0) {
+		this->m_deletionQueue.emplace_back(handle);
+	}
+}
+
+
+const Hush::RefCounted& Hush::ResourceManager::GetRefCount(const Hush::HandleId& handle) {
+	return this->m_references[handle];
+}
